requirements.txt
asgiref==3.6.0
certifi==2024.6.2
charset-normalizer==3.3.2
django==4.1.13
django-browser-reload==1.12.1
django-js-asset==2.2.0
gunicorn==22.0.0
idna==3.7
Pillow==10.3.0
psycopg2-binary==2.9.5
python-decouple==3.6
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
django-ckeditor==6.7.1
unidecode==1.3.8